working on enums
